Test doc examples
