+++ CHANGELOG
@@ -5,11 +5,10 @@ History:
 dev    2017/xx/xx
 
 
-3.0.0  2017/07/05
-      - big refactor, introduce ScreenShot class
+3.0.0  2017/07/06
+      - big refactor, introducing the ScreenShot class
       - add Numpy array interface support
-      - tests: add FPS benchmark to compare to PIL.ImageGrab
-      - examples: add OpenCV/Numpy
+      - examples: add OpenCV/Numpy, PIL pixels, FPS
 
 2.0.22  2017/04/29
       - new contributors: David Becker, redodo
