+++ docs/source/examples/linux_display_keyword.py
@@ -1,0 +1,14 @@
+# coding: utf-8
+"""
+This is part of the MSS Python's module.
+Source: https://github.com/BoboTiG/python-mss
+
+Usage example with a specific display.
+"""
+
+from mss.linux import MSS
+
+
+with MSS(display=':0.0') as sct:
+    for filename in sct.save(output=output):
+        print(filename)
