+++ docs/source/examples/part_of_screen.py
@@ -1,0 +1,23 @@
+# coding: utf-8
+"""
+This is part of the MSS Python's module.
+Source: https://github.com/BoboTiG/python-mss
+
+Example to capture part of the screen.
+"""
+
+import mss
+import mss.tools
+
+
+with mss.mss() as sct:
+    # The screen part to capture
+    monitor = {'top': 160, 'left': 160, 'width': 160, 'height': 135}
+    output = 'sct-{top}x{left}_{width}x{height}.png'.format(**monitor)
+
+    # Grab the data
+    sct_img = sct.grab(monitor)
+
+    # Save to the picture file
+    mss.tools.to_png(sct_img.rgb, sct_img.size, output)
+    print(output)
